move GetAppInfo to public controller
