+++ controllers/v1/public_controller/app_info.go
@@ -1,4 +1,4 @@
-package v1
+package public_controller
 
 import (
 	"flop/config/database"
